refactor: move global commands in global folder
